+++ CHANGES.rst
@@ -5,7 +5,7 @@ Changelog
 master
 ======
 
-- improve loading application module: allow relative imports before ``configure`` call
+- add support function-based commands
 
 1.2.0 - 2016-10-08
 ==================